+++ src/components/design-system/primitives/card.test.tsx
@@ -1,0 +1,4 @@
+import { runSnapshotTests } from '@/test/storybook-snapshots'
+import * as stories from './card.stories'
+
+runSnapshotTests(stories)