+++ vitest.setup.ts
@@ -2,3 +2,10 @@
 
 // Load .env files
 import 'dotenv/config'
+
+// Setup Storybook project annotations for portable stories (snapshot tests)
+import * as a11yAddonAnnotations from '@storybook/addon-a11y/preview'
+import { setProjectAnnotations } from '@storybook/react'
+import * as projectAnnotations from './.storybook/preview'
+
+setProjectAnnotations([a11yAddonAnnotations, projectAnnotations])